Adding convenience links to observation view
I find myself often bouncing from an observation to the station
that made the recording, so added a link to the station in the
station name above the data payload view. In addition, made a
link to db.satnogs.org for the satellite, opening a new window
as this links out to a new site.