Implement a configurable minimum horizon in network
This commit adds a "horizon" configuration item to stations. This
allows station owners to set a minimum horizon to avoid a noisy
or obstructed floor level.

The default minimum horizon is set to 10, which is still fairly
low for a satellite pass that could be captured but given the
appropriate setup someone may be successful setting it lower.

The horizon field is honored in both the calculation of upcoming
passes in the station view as well as excluding any "below horizon"
passes in prediction_windows.

In addition, the db migration will set a minimum horizon of "10"
for all existing stations in the network. This is the horizon that
was hard-coded for the upcoming passes view so the only change to
the end user will be the behavior of window prediction matching
the upcoming passes (along with the ability to configure their
minimum horizon, of course).