Force debug toolbar to show in dev
Django debug toolbar relies on INTERNAL_IPS for access control when
running in dev. This is difficult to match up in a virtualized and
dynamic environment (such as our instructions to run in docker).

This commit works around the IP check and forces the debug toolbar
to be active in dev, assuming that most dev is done in a local and
safe environment anyway.