Fix for duplicate entries in upcoming observations
Due to a bug in pyephem when a satellite is overhead it will be
returned as a "next_pass" but with a mix of data between the
current and true next passes. This change tosses out any pass
that is currently overhead from being listed in the "upcoming
observations" list of station view. This should have no adverse
impact to the user as we are assuming the list to be "upcoming"
and not include what is currently overhead, nor would SatNogs be
able to schedule anything currently overhead.

fixes satnogs/satnogs-network#199