Switches to filter observations in the observations list
This commit adds a btn-toggle class which can be applied to a
btn-group to make a toggle switch (commonly "ON/OFF") in the
btn_toggle.js file.

This new class is implemented in the observations list to provide
a set of toggles to toggle the following:

Bad observations
Unvetted observations
Good observations

These controls are themselves collapsed by default, requiring a
click to open them up.