Fix filter collapse on selected satellite
When you open an observations list after selecting a satellite
from the filter list, the filter will remain uncollapsed.
However, it is missing the "in" class, so this causes the need
for 2 clicks to attempt to collapse it, the second will not
cause the filter div to actually hide. This commit checks for
the selection and adds this class if it exists.

fixes satnogs/satnogs-network#238