Adding satellite modal links everywhere
The satellite modal that was tested in the observation view works
great! This commit adds the similar link everywhere else a
satellite is listed in a table so we have a consistent UX.

In the observation view there is only one satellite referenced,
making the modal a 1:1 mapping. As there are multiple satellites
in the table view we pull this modal in multiple times. To cut
down on possible load times getting out of hand I'm limiting the
number of rows you'll see in the home page and observations view.

I don't quite feel this iterative import approach is the best.
Feel free to decline this if there is a better way (.js it?)