Fix scheduling from satellite view
Fixes a condition when trying to schedule an observation from the satellite modal - obs_filter is triggered but there is no ground station specified so that is passed on as 'undefined', where django expects an int to filter on. With this change, if there is no ground_station offered in obs_filter it is skipped. Now all new observation calculation methods are working again.