Add pass details to payload frame
Stores the rising azimuth, max altitude, and setting azimuth for each payload scheduled in an observation when an observation is scheduled. This will help in comparing passes (ie: high pass vs low pass, obstructions, etc).

The view will not display anything if this data is missing, so legacy entries in the database that do not have this data will still display fine.

Fixes satnogs/satnogs-network#278